Refactor: Move the SynchMempoolTask into its own file.
